requirements.txt
discord.py>=2.0.0
python-dotenv>=0.21.0
aiohttp>=3.8.1
asyncio>=3.4.3
pyyaml>=6.0
pytz>=2022.7.1 